reorganize aiopsl project structure
Simplify the AIOPSL project structure by having a generic AIOPSL
project which contains build configs for each supported platform
(LS1088A, LS2085A, LS2088A).

In order to achieve this generic API and Implementation header
files have been created pointing to the correct files based on
platform defines. This allows us to simplify the project includes
so that they're now generic.

The project directory structure now also more closely mirrors the
actual files in the repository, with files included/excluded in
the build based on the build configuration.

Due to switching to build configurations instead of standalone
projects the aiopsl.a file is now built in different folders
as opposed to a generic out folder (Eclipse limitation - build
configs can't share the same output folder). This change in turn
requires modifying the associated apps and tests to link with the
new library.

The updating of the apps and tests is a good example of the minimal
changes required in order to transition to the new project system.
Only the path to the aiop_sl project and library have been modified,
as well as pre-selecting the platform for which aiopsl will be built.

For example, if the user imports the LS2088A app_process_packet then
when building this CW will automatically know to build the LS2088A
version of AIOPSL.

Signed-off-by: Serban Constantin <serban.constantin@nxp.com>