TMan : USER_OPAQUE field on 22 bits
The USER_OPAQUE field in the timer create command and in timer expiration
and TMI delete callbacks was treated as a 16b value. Its length is of 22b.
This patch writes/read using the provided 16b value to/from the correct
field position.
Note: TODO comment was added in code to remember the fix is partial.
32b values should be provided for the USER_OPAQUE variables.

Signed-off-by: Grigore Ion <ion.grigore@freescale.com>