cdma: cdma_read revert behavior of using cache disable by default
added function cdma read with no cache for accessing external
memory. Currently function is not exposed to users. The default
behavior of cdma_read was reverted so that it uses internal cache
for increased performance on datapath processing. This function
(cdma_read) cannot be used now to fetch data written by STE or FDMA
or other accelerators

Signed-off-by: Mihai Barbulescu <mihai.barbulescu@nxp.com>