repo: add .cwGeneratedFileSetLog to .gitignore
Add the .cwGeneratedFileSetLog to .gitignore because it is
always generated and doesn't contain any useful information.

The file was used by CW for MCU's Change Wizard, which allowed
to re-targed the project to a new microcontroller family. This
wizard is not even available in the networking family of CW IDEs.

Signed-off-by: Serban Constantin <serban.constantin@nxp.com>