IPSec : Race condition on forced expiration
A lifetime timer may be re-created in the timer expiration task. If the
forced expiration is requested in the same time, timer may be re-created,
although its forced expiration was requested.
In order to inform the application the timers were expired as a result of
an explicit forced expiration request, new expiration indicators were
added.

Signed-off-by: Grigore Ion <ion.grigore@freescale.com>