ipr: IPR_SW_TABLE
IPR module uses software hash tables instead of CTLU ones.

The table lines are created on demand.
With the default values a table has 128kB and a line 512 bytes.
The table lines are obtained from SLAB so a new DPBP is needed for IPR.

The lookup key is SW extracted. From it a hash is generated
and reduced to a line number.

We protect the IPR resources with OSM transitions in exclusive mode.
scope_id (uint32_t) has a portion (0x3F) reserved for OSM infrastructure.
We could use virtual addresses for scope_id since they are unique.
But the virtual addresses must be aligned to 64 to have the reserve
portion of the scope_id zero.

Signed-off-by: George Stefan <George.Stefan@nxp.com>