gso: Add missing initialization in context for IPv6
The IP version indicator in the GSO context data was left
uninitialized in case of IPv6 protocol. This added a risk of
illegaly accessing the "outer_ipv4_ptr" if the user was incorrectly
setting this to IPv4. An initialization of this attribute was added
to match the processing of the IPv4 protocol and make sure the
proper IPv6 data structure is used.

Signed-off-by: Marian Chereji <marian.chereji@nxp.com>