IPReass : Wrong reference counter address
Address of reference counter in the workspace was defined as :
	#define REF_COUNT_ADDR_DUMMY	HWC_ACC_OUT_ADDRESS +
					CDMA_REF_CNT_OFFSET
and used as an uint32_t pointer like this :
	(uint32_t *)REF_COUNT_ADDR_DUMMY
The resulted computed address is 0x30 + 0x10 = 0x40 not 0x34 as the
intention was.
The 0x40 is the address of ADC and at 0x40 offset FQD context is stored.
Unpredictable behaviour happens when reference counter is incremented or
decremented : accelerator FQID, AUC, VA2 and VA3 fields are affected.
As a result, for example the encryption of a reassembled packet fails,
when SEC accelerator is working in the buffer reuse mode. AUC is wrongly
set to 1, but the buffer reuse mode works only if AUC is 0 or 3.

Signed-off-by: Grigore Ion <ion.grigore@freescale.com>