IPReass : Configurable fragments preservation
IPR_MODE_DO_NOT_PRESERVE_FRAGS IPReass instance configuration flag
was added in order to control if the Start of Fragment bit is
set when fragments are reassembled.
This flag must be set if the reassembled packet is forwarded to the
SEC engine for encryption/decryption.
This must not be set if the applications, in further processing,
need to split the reassembled packet in the original fragments.

Signed-off-by: Grigore Ion <ion.grigore@freescale.com>