IPF_demo: initial presentation offset should be 0
IPF application is designed to work with the initial presentation segment offset set to 0. However, in the ipf_demo.c file there is an example of how to use in case the segment offset is set to a known header of set (IPv4 header offset). If from any reasons the segment offset has a different value, in the application using IPF the frame should be represented before calling ipf_generate_frag using the segment offset 0.

Signed-off-by: Chiculita Alina-nxf36786 <nxf36786@freescale.com>