IPR: Fixed IPv6 atomic fragment detection
Enabled AD bit of PAR_CFG register to allow IPv6 atomic fragments to be
marked as fragments by the parser.
Confirm that a fragment is atomic by checking the offset_and_flags field
of the ipv6fraghdr struct provided by the parser.
No security risk is involved, as we do not allocate any reassembly
resources for atomic fragments - they must be treated separately by the
apps.

Signed-off-by: Cristian-Andrei Sandu <cristian-andrei.sandu@nxp.com>