aiopsl: improve boot error handling
This patch implements a cleaner and consistent error handling scheme
for aiopsl boot.

As of now, depending on the section of code where an error may occur,
the cores either return with error or remain stuck while waiting for
the system barrier.

This patch makes the AIOP behave consistent during the boot sequence
in case of error: the first core that detects an error writes it in
both the ACGPR register and a global error var and then proceeds to
return with the corresponding error code. The global error variable
makes it possible to detect that another core (usually the master
core that provides most of the initialization) has reported an error
and is shutting down. In such case it should stop execution as well.

All the cores before returning (w/o an error) write to the ABCR reg
that the corresponding core has finished booting.

This error scheme makes it easy for MC to check and report the status
of the AIOP booting process: it waits for the ABCR register to be
written by all the cores and then checks for the ACGPR reg to see if
AIOP has finished booting with error.

Signed-off-by: Radu Alexe <radu.alexe@nxp.com>