aiopsl: improved exception handling for e200 booke critical & non-critical interrupts
Implemented a proper mechanism for handling e200 booke critical and non-critical interrupts which allows for
correct recovery of the call stack upon recovery.
A refactoring was done to also make the code for handling the various non-critical interrupts more compact
due to the fact that they share the same set of registers among them, for providing debug info
and for the restoration process.
The possibility for the user to further expand on the exception handler was also added, by overriding the
"booke_debug_hook" weak symbol; AIOPSL must also be compiled with the AIOPSL_KERNEL_EXCEPTION_HOOK symbol
defined.
The default behaviour of the debug hook is to execute an infinite loop in order for the developer
to attach to the application, prior to returning from interrupt context.

Signed-off-by: Rosioru Dragos <dragos.rosioru@nxp.com>